+++ dune/stuff/common/tuple.hh
@@ -22,7 +22,7 @@
 #define TMAX(t_, no_) (Dune::tuple_size<t_>::value >= (no_ + 1) ? no_ : 0)
 #define TELE(t_, s_, no_) typename Dune::tuple_element<TMAX(t_, no_), t_>::type::s_
 
-template <class T1, class T2, class T3, class T4, class T5, class T6, class T7, class T8, class T9, int Size>
+template <class T1, class T2, class T3, class T4, class T5, class T6, class T7, class T8, class T9, size_t Size>
 struct RightTrimTuple
 {
   typedef Dune::tuple<T1, T2, T3, T4, T5, T6, T7, T8, T9> type;
